+++ Helper/KillboardHelper.php
@@ -99,6 +99,17 @@ class KillboardHelper {
 						.				$this->getVictimImage($killmail->victim)
 						. '			</a>'
 						. '		</figure>'
+						. '		<div class="eve-online-killboard-widget-pilot-information hidden-xs hidden-lg visible-xl">'
+						. '			<span class="victimShipImage">'
+						.				$this->getVictimShipImage($killmail->victim, 32)
+						. '			</span>'
+						. '			<span class="victimCorpImage">'
+						.				$this->getVictimCorpImage($killmail->victim, 32)
+						. '			</span>'
+						. '			<span class="victimAllianceImage">'
+						.				$this->getVictimAllianceImage($killmail->victim, 32)
+						. '			</span>'
+						. '		</div>'
 						. '	</div>'
 						. '	<div class="col-xs-8 col-sm-12 col-md-12 col-lg-7">'
 						. '		<ul>'
@@ -186,6 +197,54 @@ class KillboardHelper {
 				$victimImage = EveApiHelper::getInstance()->getCharacterImageById($victimData->characterID, false, $size);
 				break;
 		} // END switch($victimData->characterID)
+
+		return $victimImage;
+	} // END public function getVictimImage(\stdClass $victimData, $size = 256)
+
+	public function getVictimCorpImage(\stdClass $victimData, $size = 256) {
+		$victimImage = null;
+
+		switch($victimData->corporationID) {
+			case 0:
+				$victimImage = null;
+				break;
+
+			default:
+				$victimImage = EveApiHelper::getInstance()->getCorporationImageById($victimData->corporationID, false, $size);
+				break;
+		} // END switch($victimData->corporationID)
+
+		return $victimImage;
+	} // END public function getVictimImage(\stdClass $victimData, $size = 256)
+
+	public function getVictimShipImage(\stdClass $victimData, $size = 256) {
+		$victimImage = null;
+
+		switch($victimData->shipTypeID) {
+			case 0:
+				$victimImage = null;
+				break;
+
+			default:
+				$victimImage = EveApiHelper::getInstance()->getShipImageById($victimData->shipTypeID, false, $size);
+				break;
+		} // END switch($victimData->shipTypeID)
+
+		return $victimImage;
+	} // END public function getVictimImage(\stdClass $victimData, $size = 256)
+
+	public function getVictimAllianceImage(\stdClass $victimData, $size = 128) {
+		$victimImage = null;
+
+		switch($victimData->allianceID) {
+			case 0:
+				$victimImage = null;
+				break;
+
+			default:
+				$victimImage = EveApiHelper::getInstance()->getAllianceImageById($victimData->allianceID, false, $size);
+				break;
+		} // END switch($victimData->allianceID)
 
 		return $victimImage;
 	} // END public function getVictimImage(\stdClass $victimData, $size = 256)