+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [In Development](https://github.com/ppfeufer/eve-online-killboard-widget)
 [Full Changelog](https://github.com/ppfeufer/eve-online-killboard-widget/compare/v0.2...HEAD)
+### Added
+- Ship, corp and alliance images
+
 ### Changed
 - Prevent direct access to our classes
 - Writing empty index.php files into our cache directories, so the directory listing doesn't work there